Remove original waxeye.coffee/js and replace with new, non-recursive implementation.
